exam prep task all done, except unit tests

+++ c-sharp-high-quality-code-two/exam-prep/ConsoleApplication3 - 28-04-2017/SchoolManagementSystem/Interfaces/IMark.cs
@@ -2,7 +2,7 @@
 {
     public interface IMark
     {
-        int MarkValue { get; }
+        float MarkValue { get; }
 
         Subject MarkSubject { get; }
     }
